Remove old analyzer views and simplify window state
