+++ Packetry/NetworkInspector/NetworkInspectorUIUtilities.swift
@@ -1,0 +1,20 @@
+import SwiftUI
+
+extension View {
+    @ViewBuilder
+    func packetryToolbarButtonStyle(prominent: Bool = false) -> some View {
+        if #available(macOS 26.0, *) {
+            if prominent {
+                self.buttonStyle(.glassProminent)
+            } else {
+                self.buttonStyle(.glass)
+            }
+        } else {
+            if prominent {
+                self.buttonStyle(.borderedProminent)
+            } else {
+                self.buttonStyle(.bordered)
+            }
+        }
+    }
+}
